test(alert): test default onDismiss
affects: @crave/farmblocks-alert

ISSUES CLOSED: #50
